requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
huggingface_hub>=0.20.0
python-docx>=1.0.0
PyMuPDF>=1.24.0

# Transformers backend (default)
torch>=2.0.0
transformers>=4.50.0
accelerate>=0.25.0
qwen-vl-utils>=0.0.8
torchvision>=0.15.0

# llama.cpp backend (optional — uncomment to enable)
# llama-cpp-python>=0.3.0
